Refactor classes: new package names & new class names
