+++ app/src/main/java/de/pscom/pietsmiet/repository/FirebaseRepository.java
@@ -1,4 +1,4 @@
-package de.pscom.pietsmiet.presenter;
+package de.pscom.pietsmiet.repository;
 
 import android.content.Context;
 
@@ -22,11 +22,11 @@ import static de.pscom.pietsmiet.util.PostType.PIETCAST;
 import static de.pscom.pietsmiet.util.PostType.PS_VIDEO;
 import static de.pscom.pietsmiet.util.PostType.UPLOADPLAN;
 
-public class FirebasePresenter extends MainPresenter {
+public class FirebaseRepository extends MainRepository {
     private static final String FIREBASE_URL = "https://pietsmiet-de5ff.firebaseio.com";
     FirebaseApiInterface apiInterface;
 
-    public FirebasePresenter(Context context) {
+    public FirebaseRepository(Context context) {
         super(context);
         RxJavaCallAdapterFactory rxAdapter = RxJavaCallAdapterFactory.createWithScheduler(Schedulers.io());
