+++ app/src/test/java/de/pscom/pietsmiet/util/PostPresenterTest.java
@@ -5,22 +5,23 @@ import org.mockito.Mock;
 
 import java.util.Date;
 
-import de.pscom.pietsmiet.MainActivity;
+import de.pscom.pietsmiet.view.MainActivity;
 import de.pscom.pietsmiet.generic.Post;
 import de.pscom.pietsmiet.presenter.FacebookPresenterTest;
 import de.pscom.pietsmiet.presenter.FirebasePresenterTest;
+import de.pscom.pietsmiet.presenter.PostPresenter;
 import de.pscom.pietsmiet.presenter.TwitterPresenterTest;
 import de.pscom.pietsmiet.presenter.YoutubePresenterTest;
 import rx.Observable;
 import rx.observers.TestSubscriber;
 
-public class PostManagerTest {
+public class PostPresenterTest {
     @Mock
     MainActivity mainActivity;
 
     @Test
     public void buildAndSortPosts() throws Exception {
-        PostManager postManager = new PostManager(mainActivity);
+        PostPresenter postPresenter = new PostPresenter(mainActivity);
         TestSubscriber<Post> testSubscriber = new TestSubscriber<>();
         Observable<Post.PostBuilder> youtubeObs = new YoutubePresenterTest().preparePresenter().fetchPostsSinceObservable(new Date());
         Observable<Post.PostBuilder> facebookObs = new FacebookPresenterTest().preparePresenter().fetchPostsSinceObservable(new Date());
